Can get all orders in service

+++ WebApplication2/Domain/Order.cs
@@ -29,7 +29,7 @@
             return totalPrice;
         }
 
-        public double getClientId()
+        public int getClientId()
         {
             return clientId;
         }
